requirements.txt
Authlib==1.3.0
blinker==1.8.2
byteplus-sdk==0.0.9
click==8.1.7
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
SQLAlchemy==2.0.30
Werkzeug==3.0.3